detail 패키지 추가 기존 2차detail패키지이름 detail2로 수정

+++ src/main/java/com/green/winey_final/detail/model/SelCountVo.java
@@ -3,8 +3,9 @@ package com.green.winey_final.detail.model;
 import lombok.Data;
 
 @Data
+
 public class SelCountVo {
     private Long productId;
-    private int reviewLevel;
+    private Long reviewLevel;
 
 }
